Create Joomla_find_plesk.sh
This script uses the Plesk command line utility plesk to get a list of all subscriptions on the server and their corresponding paths. It then checks if each site path exists and has the version.php file in the includes directory under the httpdocs directory. If the version.php file exists, it extracts the Joomla version from it and outputs the Joomla version and site name.

You can run this script on the command line of the Plesk server as the root user or a user with sudo privileges.

Here are the steps to run the script:
	1.	Open a terminal window on the Plesk server.
	2.	Create a new file for the script by typing nano script.sh (you can replace script.sh with any filename you like).
	3.	Copy and paste the script into the file and save the file by pressing Ctrl+O and then Ctrl+X.
	4.	Make the file executable by typing chmod +x script.sh.
	5.	Run the script by typing ./script.sh.

The script should output the Joomla version and site name for each Joomla installation on the server.